requirements.txt
aiofiles==24.1.0
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.6.0
Authlib==0.5.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
fastapi-mail==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
iso8601==1.1.0
Jinja2==3.1.4
mailersend==0.5.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pillow==10.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
tortoise-orm==0.21.6
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
watchfiles==0.24.0
websockets==13.1
